Devplan lib small refactoring
