--- app/models/group.rb
@@ -1,5 +1,0 @@
-class Group
-  include ActiveModel::Model
-
-  attr_accessor :id, :name, :type, :type_name
-end
