--- app/models/hour.rb
@@ -1,5 +1,0 @@
-class Hour
-  include ActiveModel::Model
-
-  attr_accessor :starts_at, :ends_at
-end
