--- app/models/room.rb
@@ -1,5 +1,0 @@
-class Room
-  include ActiveModel::Model
-
-  attr_accessor :id, :name, :full_name, :regular
-end
