+++ config/initializers/devplan.rb
@@ -1,0 +1,1 @@
+Perret::Application.config.x.devplan_url = 'http://cash.knp-dev.org/api'
