+++ lib/devplan.rb
@@ -1,0 +1,1 @@
+require 'devplan/api'
